requirements.txt
cryptography==45.0.6
fastapi==0.116.1
firebase_admin==7.1.0
grpcio-status==1.74.0
h2==4.2.0
httptools==0.6.4
openai==1.99.6
python-dotenv==1.1.1
PyYAML==6.0.2
uvicorn==0.35.0
uvloop==0.21.0
websockets==15.0.1
